requirements.txt
PyYAML
pydantic
requests
pillow
ffmpeg-python
m3u8
pysftp
elasticsearch
aiohttp
aiopath
aiofiles
sqlalchemy[asyncio]==1.4
httpx
setuptools
tweepy[async]
aiocsv
asyncpg
fastapi
numpy
starlette
sshtunnel
aiopath
youtube-dl
elasticsearch
greenlet
